requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
PyMySQL==1.1.0
google-generativeai==0.3.2
cryptography==41.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
